Remove dead toposort SQL-string sorter
The toposort package had two parallel implementations:

  1. Schema-object-based: OrderFromSchema (schema.go), the only path
     reached from production via diff_all.go.
  2. Raw-SQL-based: ExtractDeps + SortSQL (deps.go + sort.go), an
     alternative entry point that parses CREATE statements out of a
     SQL string. Nothing in the production binary calls it.

`golang.org/x/tools/cmd/deadcode -test=false ./cmd/pist` confirmed
that ~400 lines in deps.go and sort.go (and the 500+ lines of tests
that exercised them) were unreachable from the CLI entry point and
existed only because tests referenced them.

Drop deps.go, sort.go, and the corresponding tests. Move the one
function shared with the production path — qualifyRangeVar — into
schema.go, its only remaining caller. Trim the export_test.go alias
list to just NewGraph (the graph type is still used by tests of the
production path).
